some refactoring and tidy up; allow the optimized HTML rendering to be passed on the creation of the RTV
